backend/pyproject.toml
[project]
name = "photography-portfolio"
version = "0.1.0"
description = "A photography portfolio backend built with Flask"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "Flask>=3.0.0",
    "Flask-CORS>=4.0.0",
    "Pillow>=10.1.0",
    "python-dotenv>=1.0.0",
    "boto3>=1.34.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
